backend/requirements.txt
tensorflow>=2.0 # for CPU supportW
#tensorflow[and-cuda] # for GPU support
numpy
matplotlib
Pillow
scipy
opencv-python
fastapi
uvicorn
python-multipart